fixed bug in time propagation
the convolution integrals for the bold propagators are still performed with loops over time and are therefore very slow, but they do not refer to zero values for the self energy and the propagators anymore. The initial hybridization function is remained the same, so the results for the bold propagators should coincide with the time-independent code in the master branch. They are almost identical, the difference comes from the fact that i do not use the trapezoidal rule for the summations. I will incorporate this in the next step, where i perform the integrals with matrix multiplications and the np.trapez integration, which work faster.